ISSUE_10_rules: see description
Rather than having a wrapper, I decided to go for an if check.
It makes more sense like this.

It makes no sense to call a function with a global. I got
rid of the second argument.

weight_rule was broken.

Implemented new checks to see whether we should run
the extra checks which are fixable.